backend/requirements.txt
Flask==3.1.0
Flask-CORS==4.0.1
pytz==2024.2
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.16.0
webdriver-manager==4.0.2
PyPDF2==3.0.1
pdfplumber==0.11.0
openai==1.52.0
httpx==0.27.2
python-dotenv==1.0.1
schedule==1.2.1
pillow==10.4.0
sqlalchemy==2.0.35
APScheduler==3.10.4
watchdog==3.0.0


